+++ lib/carin_for_blue_button_test_kit/custom_groups/v2.0.0-dev-nonfinancial/c4bb_smart_launch_group.rb
@@ -22,15 +22,32 @@ module CarinForBlueButtonTestKit
         They then perform a standalone launch to obtain an access token which
         can be used by the remaining tests to access patient data.
       )
-      input_order :url,
-                  :standalone_client_id,
-                  :standalone_client_secret,
-                  :standalone_requested_scopes,
-                  :use_pkce,
-                  :pkce_code_challenge_method,
-                  :standalone_authorization_method,
-                  :client_auth_type,
-                  :client_auth_encryption_method
+
+      config(
+        inputs: {
+          received_scopes: { name: :standalone_received_scopes },
+          smart_auth_info: {
+            options: {
+              components: [
+                {
+                  name: :requested_scopes,
+                  default: %(
+                    launch/patient openid fhirUser
+                    patient/ExplanationOfBenefit.read patient/Coverage.read
+                    patient/Patient.read patient/Organization.read
+                    patient/Practitioner.read user/ExplanationOfBenefit.read
+                    user/Coverage.read user/Patient.read
+                    user/Organization.read user/Practitioner.read
+                  ).gsub(/\s{2,}/, ' ').strip
+                }
+              ]
+            }
+          }
+        },
+        outputs: {
+          patient_id: { name: :patient_ids }
+        }
+      )
 
       group from: :smart_discovery do
         run_as_group
@@ -79,31 +96,8 @@ module CarinForBlueButtonTestKit
           * [Standalone Launch Sequence](https://www.hl7.org/fhir/smart-app-launch/1.0.0/index.html#standalone-launch-sequence)
         )
 
-        config(
-          inputs: {
-            requested_scopes: {
-              default: %(
-                launch/patient openid fhirUser
-                patient/ExplanationOfBenefit.read patient/Coverage.read
-                patient/Patient.read patient/Organization.read
-                patient/Practitioner.read user/ExplanationOfBenefit.read
-                user/Coverage.read user/Patient.read
-                user/Organization.read user/Practitioner.read
-              ).gsub(/\s{2,}/, ' ').strip
-            }
-          },
-          outputs: {
-            patient_id: { name: :patient_ids },
-            smart_credentials: { name: :smart_credentials }
-          }
-        )
-
         test from: :c4bb_v200devnonfinancial_smart_scopes do
           config(
-            inputs: {
-              requested_scopes: { name: :standalone_requested_scopes },
-              received_scopes: { name: :standalone_received_scopes }
-            },
             options: {
               required_scopes: %w[
                 openid