+++ spec/carin_for_blue_button/c4bb_client/c4bb_client_eob_required_searches_spec.rb
@@ -2,11 +2,8 @@ require_relative '../../../lib/carin_for_blue_button_test_kit/client/v2.0.0/clai
                  'patient_claims_data_request_test'
 
 RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
-  let(:suite) { Inferno::Repositories::TestSuites.new.find('c4bb_v200_client') }
+  let(:suite_id) { 'c4bb_v200_client' }
   let(:test) { Inferno::Repositories::Tests.new.find('eob_required_searches') }
-  let(:session_data_repo) { Inferno::Repositories::SessionData.new }
-  let(:results_repo) { Inferno::Repositories::Results.new }
-  let(:test_session) { repo_create(:test_session, test_suite_id: 'c4bb_v200_client') }
   let(:result) { repo_create(:result, test_session_id: test_session.id) }
 
   let(:c4bb_eob_include_bundle) do
@@ -62,7 +59,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
      { eob_id_search => [eob_id_search_tags, c4bb_eob_search_bundle] }]
   end
   let(:eob_include_search_tags) do
-    ['carin_resource_api',
+    ['resource_api',
      'ExplanationOfBenefitSearch',
      'ExplanationOfBenefit_Inpatient_Institutional',
      'ExplanationOfBenefit_Outpatient_Institutional',
@@ -84,7 +81,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
   end
 
   let(:eob_include_all_search_tags) do
-    ['carin_resource_api',
+    ['resource_api',
      'ExplanationOfBenefitSearch',
      'ExplanationOfBenefit_Inpatient_Institutional',
      'ExplanationOfBenefit_Outpatient_Institutional',
@@ -102,7 +99,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
 
   let(:eob_required_search_tags) do
     [
-      'carin_resource_api',
+      'resource_api',
       'ExplanationOfBenefitSearch',
       'ExplanationOfBenefit_Inpatient_Institutional',
       'identifier',
@@ -115,7 +112,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
 
   let(:eob_last_upated_search_tags) do
     [
-      'carin_resource_api',
+      'resource_api',
       'ExplanationOfBenefitSearch',
       'ExplanationOfBenefit_Inpatient_Institutional',
       '_lastUpdated'
@@ -124,7 +121,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
 
   let(:eob_id_search_tags) do
     [
-      'carin_resource_api',
+      'resource_api',
       'ExplanationOfBenefitSearch',
       'ExplanationOfBenefit_Inpatient_Institutional',
       '_id'
@@ -162,20 +159,6 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientEOBRequiredSearches do
       headers:,
       tags:
     )
-  end
-
-  def run(runnable, inputs = {})
-    test_run_params = { test_session_id: test_session.id }.merge(runnable.reference_hash)
-    test_run = Inferno::Repositories::TestRuns.new.create(test_run_params)
-    inputs.each do |name, value|
-      session_data_repo.save(
-        test_session_id: test_session.id,
-        name:,
-        value:,
-        type: runnable.config.input_type(name) || 'text'
-      )
-    end
-    Inferno::TestRunner.new(test_session:, test_run:).run(runnable)
   end
 
   it 'passes if successful FHIR search requests for all required searches for EOB was made' do