+++ spec/carin_for_blue_button/c4bb_client/c4bb_client_initial_wait_test_spec.rb
@@ -5,11 +5,9 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
   include Rack::Test::Methods
   include RequestHelpers
 
-  let(:suite) { Inferno::Repositories::TestSuites.new.find('c4bb_v200_client') }
+  let(:suite_id) { 'c4bb_v200_client' }
   let(:test) { Inferno::Repositories::Tests.new.find('c4bb_v200_initial_wait_test') }
-  let(:session_data_repo) { Inferno::Repositories::SessionData.new }
   let(:results_repo) { Inferno::Repositories::Results.new }
-  let(:test_session) { repo_create(:test_session, test_suite_id: 'c4bb_v200_client') }
 
   let(:base_url) { "#{Inferno::Application['base_url']}/custom/c4bb_v200_client/fhir" }
   let(:fhir_server) { ENV.fetch('FHIR_REFERENCE_SERVER') }
@@ -28,7 +26,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
 
   let(:reference_server_token) { 'SAMPLE_TOKEN' }
   let(:client_id) { 'SAMPLE_CLIENT_ID' }
-  let(:bearer_token) { JWT.encode({ inferno_client_id: 'SAMPLE_CLIENT_ID' }, nil, 'none') }
+  let(:bearer_token) { SMARTAppLaunch::MockSMARTServer.client_id_to_token(client_id, 5) }
 
   let(:resume_claims_data_url) do
     "#{Inferno::Application['base_url']}/custom/c4bb_v200_client/resume_claims_data?test_run_identifier=#{client_id}"
@@ -61,21 +59,6 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
     bundle
   end
 
-  def run(runnable, inputs = {})
-    test_run_params = { test_session_id: test_session.id }.merge(runnable.reference_hash)
-    test_run = Inferno::Repositories::TestRuns.new.create(test_run_params)
-
-    inputs.each do |name, value|
-      session_data_repo.save(
-        test_session_id: test_session.id,
-        name:,
-        value:,
-        type: runnable.config.input_type(name)
-      )
-    end
-    Inferno::TestRunner.new(test_session:, test_run:).run(runnable)
-  end
-
   it 'Passes and responds 200 if valid Patient API request sent to the provided URL and Bearer token is correct' do
     allow(test).to receive_messages(suite:, parent: suite)
 
@@ -97,7 +80,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
 
     expect(response_body['resourceType']).to eq('Bundle')
     expect(response_body['entry'].first['resource']['resourceType']).to eq('Patient')
-    expect(last_request.env['inferno.tags']).to include('carin_resource_api', 'Patient', '_id')
+    expect(last_request.env['inferno.tags']).to include('resource_api', 'Patient', '_id')
     get(resume_claims_data_url)
     result = results_repo.find(result.id)
     expect(result.result).to eq('pass')
@@ -126,7 +109,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
     expect(response_body['resourceType']).to eq('Bundle')
     expect(response_body['entry'].length).to eq 10
     expect(response_body['entry'].first['resource']['resourceType']).to eq('ExplanationOfBenefit')
-    expect(last_request.env['inferno.tags']).to include('carin_resource_api',
+    expect(last_request.env['inferno.tags']).to include('resource_api',
                                                         'ExplanationOfBenefit_Inpatient_Institutional',
                                                         'ExplanationOfBenefit_Outpatient_Institutional',
                                                         'ExplanationOfBenefit_Oral',
@@ -154,7 +137,7 @@ RSpec.describe CarinForBlueButtonTestKit::C4BBClientInitialWaitTest do
 
     result = run(test, client_id:)
 
-    header 'Authorization', 'Bearer WRONG_TOKEN'
+    header 'Authorization', "Bearer #{SMARTAppLaunch::MockSMARTServer.client_id_to_token('wrong_client_id', 5)}"
     get(patient_api_request)
 
     expect(last_response).to be_server_error